@@ -1,6 +1,6 @@
 {
   "name": "canon-react",
-  "version": "0.3.2",
+  "version": "0.6.4",
   "description": "Rackspace Canon components built with React",
   "repository": {
     "type": "git",
@@ -8,12 +8,10 @@
   },
   "scripts": {
     "prepublish": "./prepublish.sh",
-    "build": "./node_modules/.bin/grunt build",
+    "build": "scripts/cibuild",
     "clean": "./node_modules/.bin/grunt clean",
-    "demo-build": "./node_modules/.bin/grunt demo-build",
-    "documentation-build": "./node_modules/.bin/grunt documentation-build",
-    "test-watch": "./node_modules/.bin/grunt watch 2>&1 >/dev/null & ./node_modules/karma/bin/karma start karma.dev.js",
-    "test": "./node_modules/.bin/grunt build && ./node_modules/karma/bin/karma start karma.ci.js"
+    "test": "scripts/cibuild",
+    "documentation-build": "./node_modules/.bin/grunt documentation-build"
   },
   "keywords": [
     "react",
@@ -42,40 +40,46 @@
     "react": "global:React"
   },
   "dependencies": {
-    "babel-runtime": "^5.5.8",
-    "tether": "1.0.2"
+    "babel-plugin-transform-runtime": "6.4.3",
+    "babel-runtime": "6.3.19",
+    "classnames": "^2.2.3",
+    "tether": "^1.1.0"
   },
   "peerDependencies": {
     "react": ">=0.13.3"
   },
   "devDependencies": {
-    "babel": "^5.5.8",
-    "babel-runtime": "^5.5.8",
-    "babelify": "^6.1.2",
+    "babel": "6.3.26",
+    "babel-eslint": "4.1.6",
+    "babel-preset-es2015": "6.3.13",
+    "babel-preset-react": "6.3.13",
+    "babelify": "7.2.0",
     "browserify": "<4.0.0",
     "browserify-shim": "3.8.9",
-    "eslint": "^1.2.1",
-    "eslint-plugin-react": "^3.3.0",
+    "eslint": "^1.10.3",
+    "eslint-plugin-react": "^3.15.0",
     "grunt": "~0.4.2",
-    "grunt-babel": "^5.0.1",
+    "grunt-babel": "6.0.0",
     "grunt-browserify": "~1.3.0",
     "grunt-cli": "~0.1.13",
     "grunt-contrib-clean": "~0.5.0",
     "grunt-contrib-copy": "~0.5.0",
     "grunt-contrib-uglify": "~0.3.2",
     "grunt-contrib-watch": "~0.5.3",
-    "grunt-eslint": "^15.0.0",
-    "grunt-jsxhint": "~0.5.0",
+    "grunt-eslint": "^17.3.1",
     "grunt-shell": "~0.6.4",
-    "karma": "^0.12.22",
-    "karma-browserify": "~3.0.1",
-    "karma-chrome-launcher": "^0.1.4",
-    "karma-failed-reporter": "0.0.2",
-    "karma-jasmine": "^0.1.5",
-    "karma-junit-reporter": "^0.2.2",
-    "karma-phantomjs-launcher": "^0.1.4",
-    "karma-react-jsx-preprocessor": "0.1.1",
-    "karma-spec-reporter": "0.0.13",
+    "http-server": "^0.8.5",
+    "jasmine-core": "^2.4.1",
+    "karma": "^0.13.19",
+    "karma-browserify": "^4.4.2",
+    "karma-chrome-launcher": "^0.2.2",
+    "karma-failed-reporter": "0.0.3",
+    "karma-jasmine": "^0.3.6",
+    "karma-junit-reporter": "^0.3.8",
+    "karma-phantomjs-launcher": "^0.2.3",
+    "karma-react-jsx-preprocessor": "^0.1.1",
+    "karma-spec-reporter": "0.0.23",
+    "phantomjs": "^1.9.19",
     "react": "~0.13.2"
   }
 }
