@@ -1,6 +1,5 @@
-var React = require('react');
-
-var DemoView = require('./DemoView');
+import React from 'react';
+import DemoView  from './DemoView';
 
 (function () {
   React.render(<DemoView/>, document.getElementById('content'));
