@@ -1,16 +1,17 @@
 import React from 'react';
+import classNames from 'classnames';
 
 class ButtonGroup extends React.Component {
   render() {
     let classes;
 
-    classes = ['rs-btn-group'];
-    if (this.props.className) {
-      classes.push(this.props.className);
-    }
+    classes = classNames(
+      'rs-btn-group',
+      this.props.className
+    );
 
     return (
-      <div {...this.props} className={classes.join(' ')}>
+      <div {...this.props} className={classes}>
         {this.props.children}
       </div>
     );
