@@ -1,4 +1,5 @@
 import React from 'react';
+import classNames from 'classnames';
 
 import ButtonGroup from './ButtonGroup';
 
@@ -6,7 +7,7 @@ class PopoverFooter extends React.Component {
 
   render () {
     return (
-      <ButtonGroup className='rs-popover-footer'>
+      <ButtonGroup className={classNames('rs-popover-footer', this.props.className)}>
         {this.props.children}
       </ButtonGroup>
     );
