@@ -1,15 +1,15 @@
 import React from 'react';
+import classNames from 'classnames';
 
 class ProcessingIndicator extends React.Component {
 
   render() {
     let classes;
 
-    classes = 'rs-processing-indicator';
-
-    if (this.props.hidden) {
-      classes += ' rs-hidden';
-    }
+    classes = classNames(
+      'rs-processing-indicator',
+      { 'rs-hidden': this.props.hidden }
+    );
 
     return (
       <i className={classes}></i>
