@@ -1,6 +1,7 @@
 import React from 'react';
+import classNames from 'classnames';
 
-const SizeClasses = {
+const SIZE_CLASSES = {
   'xsmall': 'rs-progress-xsmall',
   'small': 'rs-progress-small',
   'medium': 'rs-progress-medium',
@@ -8,42 +9,37 @@ const SizeClasses = {
   'xlarge': 'rs-progress-xlarge'
 };
 
-const StatusClasses = {
+const STATUS_CLASSES = {
   'ok': 'rs-status-ok',
   'error': 'rs-status-error',
   'warning': 'rs-status-warning',
   'info': 'rs-status-info'
 };
 
-const TypeClasses = {
+const TYPE_CLASSES = {
   'solid': 'rs-bar-solid',
   'striped': 'rs-bar-striped'
 };
 
-class ProgressBar extends React.Component{
+class ProgressBar extends React.Component {
 
   _getSizeClass() {
-    var sizeClass;
-
-    sizeClass = 'rs-progress';
-
-    if (SizeClasses[this.props.size]) {
-      sizeClass += ' ' + SizeClasses[this.props.size];
-    }
-    return sizeClass;
+     return classNames(
+      'rs-progress',
+      SIZE_CLASSES[this.props.size]
+    );
   }
 
   _getStatusClass() {
-    var statusClass;
-
-    statusClass = ['rs-bar'];
-    statusClass.push(StatusClasses[this.props.status]);
-    statusClass.push(TypeClasses[this.props.type]);
-    return statusClass.join(' ');
+    return classNames(
+      'rs-bar',
+      STATUS_CLASSES[this.props.status],
+      TYPE_CLASSES[this.props.type]
+    );
   }
 
   render() {
-    var style, width;
+    let style, width;
 
     width = this.props.progress + '%';
     style = { 'width': width };
