@@ -1,5 +1,12 @@
 import Button from './Button';
 import ButtonGroup from './ButtonGroup';
+import Criteria from './Criteria';
+import Divider from './Divider';
+import Dropdown from './Dropdown';
+import DropdownItem from './DropdownItem';
+import DropdownTrigger from './DropdownTrigger';
+import Facet from './Facet';
+import FacetsSection from './FacetsSection';
 import Popover from './Popover';
 import PopoverBody from './PopoverBody';
 import PopoverFooter from './PopoverFooter';
@@ -9,15 +16,22 @@ import ProgressBar from './ProgressBar';
 import StatusIndicator from './StatusIndicator';
 import TooltipTrigger from './TooltipTrigger';
 
-export default {
-  Button: Button,
-  ButtonGroup: ButtonGroup,
-  Popover: Popover,
-  PopoverBody: PopoverBody,
-  PopoverFooter: PopoverFooter,
-  ProcessingIndicator: ProcessingIndicator,
-  PopoverOverlay: PopoverOverlay,
-  ProgressBar: ProgressBar,
-  StatusIndicator: StatusIndicator,
-  TooltipTrigger: TooltipTrigger
+export {
+  Button,
+  ButtonGroup,
+  Criteria,
+  Divider,
+  Dropdown,
+  DropdownItem,
+  DropdownTrigger,
+  Facet,
+  FacetsSection,
+  Popover,
+  PopoverBody,
+  PopoverFooter,
+  ProcessingIndicator,
+  PopoverOverlay,
+  ProgressBar,
+  StatusIndicator,
+  TooltipTrigger
 };
